+++ neocarta/ingest/rdbms/load.py
@@ -28,6 +28,7 @@ from ...data_model.rdbms import (
     Value,
 )
 from ...enums import NodeLabel, RelationshipType
+from ..indexes import create_full_text_index
 from ..utils import (
     _build_node_ingest_query,
     _build_relationship_ingest_query,
@@ -56,6 +57,12 @@ class Neo4jRDBMSLoader:
             database_name=self.database_name,
         )
 
+        self._create_full_text_index = partial(
+            create_full_text_index,
+            neo4j_driver=self.neo4j_driver,
+            database_name=self.database_name,
+        )
+
     def load_database_nodes(
         self,
         database_nodes: list[Database],
@@ -81,11 +88,14 @@ class Neo4jRDBMSLoader:
         schema_nodes: list[Schema],
         overwrite_existing: bool = False,
         properties_list: list[str] = ["name", "description"],
+        create_full_text_index: bool = True,
     ) -> dict:
         """Load Schema nodes into Neo4j."""
         _validate_properties_list(Schema, properties_list)
 
         self._write_node_constraint(node_labels=[NodeLabel.SCHEMA])
+        if create_full_text_index:
+            self._create_full_text_index(node_labels=[NodeLabel.SCHEMA])
         query = _build_node_ingest_query(NodeLabel.SCHEMA, overwrite_existing, properties_list)
 
         _, summary, _ = self.neo4j_driver.execute_query(
@@ -101,11 +111,14 @@ class Neo4jRDBMSLoader:
         table_nodes: list[Table],
         overwrite_existing: bool = False,
         properties_list: list[str] = ["name", "description"],
+        create_full_text_index: bool = True,
     ) -> dict:
         """Load Table nodes into Neo4j."""
         _validate_properties_list(Table, properties_list)
 
         self._write_node_constraint(node_labels=[NodeLabel.TABLE])
+        if create_full_text_index:
+            self._create_full_text_index(node_labels=[NodeLabel.TABLE])
         query = _build_node_ingest_query(NodeLabel.TABLE, overwrite_existing, properties_list)
 
         _, summary, _ = self.neo4j_driver.execute_query(
@@ -128,11 +141,14 @@ class Neo4jRDBMSLoader:
             "is_primary_key",
             "is_foreign_key",
         ],
+        create_full_text_index: bool = True,
     ) -> dict:
         """Load Column nodes into Neo4j."""
         _validate_properties_list(Column, properties_list)
 
         self._write_node_constraint(node_labels=[NodeLabel.COLUMN])
+        if create_full_text_index:
+            self._create_full_text_index(node_labels=[NodeLabel.COLUMN])
         query = _build_node_ingest_query(NodeLabel.COLUMN, overwrite_existing, properties_list)
 
         _, summary, _ = self.neo4j_driver.execute_query(
@@ -320,11 +336,14 @@ class Neo4jRDBMSLoader:
         business_term_nodes: list[BusinessTerm],
         overwrite_existing: bool = False,
         properties_list: list[str] = ["name", "description"],
+        create_full_text_index: bool = True,
     ) -> dict:
         """Load BusinessTerm nodes into Neo4j."""
         _validate_properties_list(BusinessTerm, properties_list)
 
         self._write_node_constraint(node_labels=[NodeLabel.BUSINESS_TERM])
+        if create_full_text_index:
+            self._create_full_text_index(node_labels=[NodeLabel.BUSINESS_TERM])
         query = _build_node_ingest_query(
             NodeLabel.BUSINESS_TERM, overwrite_existing, properties_list
         )